Updated to tensorflow 1.6 and spark 2.3.  (#139)
* Updated to tensorflow 1.5 and spark 2.3. Mostly minor changes related to Tensor taking a type parameter. Updated the protobuf files and regenerated their java sources.

* Bumped tf version in travis config to 1.5

* Travis fix, try to run on ubuntu16.04 (unsuported yet but might work?) as tf1.5 does not run on ubuntu14.04

* Fixed test, added type parameter to Tensor

* Updated to ft1.6. Removed test configurations with Spark 2.0

* fixed travis setup

* Regenerated java files from the protobuf sources updated to tf1.6

+++ src/main/java/org/tensorflow/framework/SaveSliceInfoDefOrBuilder.java
@@ -8,95 +8,95 @@ public interface SaveSliceInfoDefOrBuilder extends
     com.google.protobuf.MessageOrBuilder {
 
   /**
-   * <code>optional string full_name = 1;</code>
-   *
    * <pre>
    * Name of the full variable of which this is a slice.
    * </pre>
+   *
+   * <code>string full_name = 1;</code>
    */
   java.lang.String getFullName();
   /**
-   * <code>optional string full_name = 1;</code>
-   *
    * <pre>
    * Name of the full variable of which this is a slice.
    * </pre>
+   *
+   * <code>string full_name = 1;</code>
    */
   com.google.protobuf.ByteString
       getFullNameBytes();
 
   /**
-   * <code>repeated int32 full_shape = 2;</code>
-   *
    * <pre>
    * Shape of the full variable.
    * </pre>
+   *
+   * <code>repeated int64 full_shape = 2;</code>
    */
-  java.util.List<java.lang.Integer> getFullShapeList();
+  java.util.List<java.lang.Long> getFullShapeList();
   /**
-   * <code>repeated int32 full_shape = 2;</code>
-   *
    * <pre>
    * Shape of the full variable.
    * </pre>
+   *
+   * <code>repeated int64 full_shape = 2;</code>
    */
   int getFullShapeCount();
   /**
-   * <code>repeated int32 full_shape = 2;</code>
-   *
    * <pre>
    * Shape of the full variable.
    * </pre>
+   *
+   * <code>repeated int64 full_shape = 2;</code>
    */
-  int getFullShape(int index);
+  long getFullShape(int index);
 
   /**
-   * <code>repeated int32 var_offset = 3;</code>
-   *
    * <pre>
    * Offset of this variable into the full variable.
    * </pre>
+   *
+   * <code>repeated int64 var_offset = 3;</code>
    */
-  java.util.List<java.lang.Integer> getVarOffsetList();
+  java.util.List<java.lang.Long> getVarOffsetList();
   /**
-   * <code>repeated int32 var_offset = 3;</code>
-   *
    * <pre>
    * Offset of this variable into the full variable.
    * </pre>
+   *
+   * <code>repeated int64 var_offset = 3;</code>
    */
   int getVarOffsetCount();
   /**
-   * <code>repeated int32 var_offset = 3;</code>
-   *
    * <pre>
    * Offset of this variable into the full variable.
    * </pre>
+   *
+   * <code>repeated int64 var_offset = 3;</code>
    */
-  int getVarOffset(int index);
+  long getVarOffset(int index);
 
   /**
-   * <code>repeated int32 var_shape = 4;</code>
-   *
    * <pre>
    * Shape of this variable.
    * </pre>
+   *
+   * <code>repeated int64 var_shape = 4;</code>
    */
-  java.util.List<java.lang.Integer> getVarShapeList();
+  java.util.List<java.lang.Long> getVarShapeList();
   /**
-   * <code>repeated int32 var_shape = 4;</code>
-   *
    * <pre>
    * Shape of this variable.
    * </pre>
+   *
+   * <code>repeated int64 var_shape = 4;</code>
    */
   int getVarShapeCount();
   /**
-   * <code>repeated int32 var_shape = 4;</code>
-   *
    * <pre>
    * Shape of this variable.
    * </pre>
+   *
+   * <code>repeated int64 var_shape = 4;</code>
    */
-  int getVarShape(int index);
+  long getVarShape(int index);
 }
